Add files via upload
Star Merchant is a game I wrote over 40 years ago.  It was published in the August 1981 issue of Creative Computing.   It was later included in David Ahl's 1984 book, Big Computer Games.    I have seen several copies of it on line and discuss this further in the pdf file.   I have recently rehosted it to the TRS-80 Model 100 and NEC PC-8201A.   I tweaked a few of the parameters to try and make it easier to win.   I hope you enjoy it.   Lloyd